more class refactor
