+++ App/app/src/main/java/com/flaredown/flaredownApp/Helpers/API/API_Error.java
@@ -1,4 +1,4 @@
-package com.flaredown.flaredownApp.FlareDown;
+package com.flaredown.flaredownApp.Helpers.API;
 
 import com.android.volley.VolleyError;
