+++ App/app/src/main/java/com/flaredown/flaredownApp/Settings/FragmentTreatmentReminder.java
@@ -28,7 +28,7 @@ import com.flaredown.flaredownApp.Helpers.TimeHelper;
 import com.flaredown.flaredownApp.Models.Alarm;
 import com.flaredown.flaredownApp.R;
 import com.flaredown.flaredownApp.Receivers.AlarmReceiver;
-import com.flaredown.flaredownApp.FlareDown.Locales;
+import com.flaredown.flaredownApp.Helpers.Locales;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
