+++ App/app/src/main/java/com/flaredown/flaredownApp/Toolbars/MainToolbarView.java
@@ -1,4 +1,4 @@
-package com.flaredown.flaredownApp;
+package com.flaredown.flaredownApp.Toolbars;
 
 import android.app.Activity;
 import android.content.Context;
@@ -10,6 +10,8 @@ import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.LinearLayout;
 import android.widget.TextView;
+
+import com.flaredown.flaredownApp.R;
 
 /**
  * Created by thunter on 13/09/15.
