+++ App/app/src/main/res/layout/checkin_activity.xml
@@ -2,7 +2,8 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
     <!--<include layout="@layout/toolbar"/>-->
-    <com.flaredown.flaredownApp.MainToolbarView android:layout_width="match_parent"
+    <com.flaredown.flaredownApp.Toolbars.MainToolbarView
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:id="@+id/main_toolbar_view"
         android:layout_alignParentTop="true"
