+++ App/app/src/main/res/layout/checkin_activity_add_dialog.xml
@@ -3,10 +3,11 @@
     android:layout_height="match_parent"
     tools:context="com.flaredown.flaredownApp.Checkin.AddEditableActivity">
 
-    <com.flaredown.flaredownApp.MainToolbarView android:layout_width="match_parent"
+    <com.flaredown.flaredownApp.Toolbars.MainToolbarView
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:id="@+id/main_toolbar_view">
-    </com.flaredown.flaredownApp.MainToolbarView>
+    </com.flaredown.flaredownApp.Toolbars.MainToolbarView>
 
     <RelativeLayout
         android:layout_width="fill_parent"
