+++ App/app/src/main/res/layout/login_activity.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.flaredown.flaredownApp.LoginActivity"
+    tools:context=".Login.LoginActivity"
     tools:ignore="fontPath">
     <!-- Login progress -->
